+++ CampusNavigator/CampusNavigator/CampusNavigator/Components/BottomNavItem.swift
@@ -1,0 +1,8 @@
+//
+//  BottomNavItem.swift
+//  CampusNavigator
+//
+//  Created by Sasindu 034 on 2025-02-23.
+//
+
+import Foundation
