+++ CampusNavigator/CampusNavigator/CampusNavigator/Helpers/ColorExtensions.swift
@@ -1,0 +1,5 @@
+import SwiftUI
+
+extension Color {
+    static let customDarkGreen = Color(red: 0/255, green: 55/255, blue: 6/255)
+}
